requirements.txt
Flask
firebase-admin
python-dotenv
Werkzeug
